Bulk upload of 3 coding samples
These coding samples are from different courses taken from the University of New Hampshire.
==================================

3P
The contents of this folder include input files for the associated code that performs rotations, negatives, and filtering of PNG images. The main goal of the assignment from my Data Structures course was to filter a person out of a series of images by creating a 3-dimensional array of pixel data from a series of disturbed photos, finding the median value (without image disruption), and present the original background without obstruction.
This code sample was written in C++.

==================================

711 lab1
The files in this directory are coding examples for an assignment from my Advanced Digital Systems course. The purpose of the assignment was to implement binary-coded-decimal adders in both 4-bit and 8-bit capacities.
This code sample was written in Verilog.

==================================

MatLab_Assignment_6
The files in this directory sample some of my coursework in my Digital Signal Processing coursework. The purpose of this assignment was to demonstrate and explore concepts relating to Decimation and Interpolation as related to digital signals, particularly in reference to the included audio file.
These coding samples were written in MatLab.

==================================